Lots of changes
Need to pin the docker images as well as add a dep that was upgraded out
from under python's docker image.

I also had to specify real users for the SLUMBER work, and the real
hostname for the the env vars.

I also fixed up the docker image for rtd to have the tex work needed to
build pdfs, since that's on be default, and I was tired of it failing
builds on creation.

I also upgraded our RTD code to the latest master, since there were a
bunch of transative deps that were becoming problematic to use
unpinned/unfrozen.

Also adding 5 minute timeouts for the proxy pass

[ticket: DEVOPS-969]
[ticket: BAC-1696]