fix: usr: Correcting Version URLs
The template was dropping all the project information from the URLs for
versions in the default template. This change is what was suggested in
this issue: https://github.com/rtfd/readthedocs.org/issues/2162

It worked locally, so I'm going to make a PR against the project, and
commit this fix in the meantime so we're working again.

[ticket: DEVTOOLS-196]